Contact manipulation .php files now handle errors
The addContact.php and searchContacts.php files now check if the UserID is in use, and the deleteContact.php and modifyContact.php files now check if the UserID and ID fields match an entry in the contact database. In both instances, while bad requests would have failed regardless, they now respond with an actual error.